[Update]기존 textLabel을 활용한 로직을 tag를 활용한 방식으로 변경 #14
